Updated from live response generation to pre-test response generation
Opted to move from the format of the client making an n-xiB request and the host generating that data on the spot then replying with it.  Moved to the client sending an initial request then having the host generate this data upfront and replying with it for each following request.  This should more accurately replicate the conditions of a database already having data in-memory and the client making a request for data.

Also changed some comments to be more descriptive, moved functions around, cut out some fat.

Client:
    - Changed client request-message from reusable to a list of requests
    - Added an initial request that tells the host what data to generate upfront
    - Added handling for likely errors

Host:
    - Added a memory-check to see if host is capable of supporting the clients demands
    - Changed some one-line functions to lambdas
    - Added some pretty user updates
    - Added handling for initial request/proceeding requests